Remove any capability constraints because of the used MODE.
As we are not aware of the relation (if any) between the real DB URLs from the set, it is not correct to restrict DB capabilities based on the used MODE. Like if it is Round Robin, do not do UPDATE ever. Maybe the URLs point to the nodes of the same cluster.

(cherry picked from commit 305a570f7dc48a8bb4ffd60f33b0df3438fd0246)