dialog: add mechanism that allows tagging dialogs when replicating
Dialogs can be labeled from the the script using the set_dlg_repl_tag()
function and each tag has a state(active/backup) which can be set via
the "set_repl_tag_active" MI cmd. The tags and their states are shared
in the cluster and nodes automatically go to the backup state when
another node is set to active.