dialog: call values/profiles write callbacks before replicating dialogs
Rename DLGCB_DB_WRITE_VP to DLGCB_WRITE_VP since it should also be
triggered when replicating a created dialog. This allow modules
to properly save context information in the dialog which can then be
visibile on all instances where the dialog is replicated.

This fixes issues with received replicated dialogs in acc, cgrates and
siprec modules.