Fixed checking IP addr against host/name
Instead of coverting an IP addr to string and to string matching, it is more sane to try to covert the host/name to IP addr (anyhow in 99% it will be an IP as it is taken from VIA hdr). The old string matching fails for IPv6 as the string representation of an IPv6 is not unique (like for IPv4) - you may have compact format for IPv6.

Reported by Pasan Meemaduma

(cherry picked from commit a69f6de764cefab7cb7179b2f439780e74082461)