dialog: fix DB operations to be performed by all nodes in the cluster
In the context of dialog replication and with the new feature of
synchronization at startup, there is no longer the need for the DB backend
to be shared between the nodes in the cluster. Each node can still have a
local dialog table but only for restart persistency as a failsafe for the
binary replication.

(cherry picked from commit d11aa92bb9aff955487988d6ae564501586a5473)