Fix dialog unref for local BYEs.
Instead of unref-ing the dialog when a final reply is received for the local BYE, better wait until the BYE transaction is destroyed. IF not, we will end up with a dangling dailog pointer in the BYE transaction.
Fix for #1330 where a local BYE retransmission in combination with TH ends up acessing the dangling dialog pointer.
Many thanks to Nick Altmann for the support in investigating this issue.

(cherry picked from commit ec52613da12d43299edf0dab26ae80e5d2153854)