Fix function get_local_contact() to add transport
to have consistency over b2b and presence module, fix the get_local_contact() to also add the transport indication in the URI it builds. This will simplify the code in presence (as we do not have to add the transport param later, from several parts of the code) and also fixes the b2b code where the transport was not added at all in the URIs.

Credits go to @arttemk for identifing the problem and suggesting the fix.
Fix for  #1275

(cherry picked from commit eb50b9b1516599316752ad5a4d037d10f625bc60)