clusterer: fix learning capabilities when restarting a node
This issue occurs when a node that has all the links with its neighbors
established is restarted and boots up faster than the next ping from another
node in the cluster. That node does not detect the failed link and never sends
the capabilities information itself to the restarted node. The solution is to
require a reply to an initial capabilities update sent when a node establishes a
link.

(cherry picked from commit ca84d6c856aee5ba169237dbb56ce2fcf0fef431)