Fix memory leak when ping-failed contacts unregister.
As for the contacts with failed ping we keep the ping_cell into hash (in order to count the failures), we need a second timer list to be sure that the ping_cellis actually removed from hash as some point. Without this second list, as the ping_cell exists only in hash, if the contact is unregistered, it will never be removed.

(cherry picked from commit 8532f9ed50aef1dbbf06668dcd9cfed1702c5a7d)