dialog: fix profile counting when replicating both profiles and dialogs
A replicated dialog was also counted as local for a replicated profile
on each node in the dialog replication cluster affecting both the
returned profile size and the broadcasted counter.
The issue is fixed by using the sharing tags mechanism and only count
a dialog in a replicated profile if it is marked as active.

(cherry picked from commit 1010f2b307706fa3f374477b39755d95091a0d6a)