dialog: Fix dialog matching bugs in looping + TH scenarios
This patch adds an extra control parameter to topology_hiding_match()
and match_dialog() which allows the script writer to disable the current
behavior of forcefully imposing the "DID_FALLBACK" dialog matching mode.

In looping scenarios, performing SIP-wise dialog matching may be out of
the question, as we may randomly match one of multiple dialogs with
identical SIP coordinates.