dialog: Always match dialogs using dlg_match_mode
This patch changes the default behavior of match_dialog() and
topology_hiding_match(), where they would both forcibly use DID_FALLBACK
matching, regardless of the value of the "dlg_match_mode" modparam.

Starting with 2.4, we change this behavior such that they default
to simply playing along with "dlg_match_mode", which is a lot
more intuitive from a script writing perspective.

If need be, a specific dialog matching behavior may be enforced for
certain types of traffic using the newly added optional parameters of
match_dialog() and topology_hiding_match().