tls_mgm: support SNI for tls server domains
Support TLS Server Name Indication in order to present the proper certificate
when hosting multiple domains on the same IP.
If the hostname that the client attempts to connect to is supplied, that
hostname is then matched against the name of the defined virtual TLS domains.
The certificate and settings of that domain are further used for the TLS
handshake with the client.