core: Greatly reduce chance of truncated corefiles
Before commit 8073d4de8ed, the SIGTERM signal delivered by the attendant
process to a process who had already crashed and would be dumping its
memory inside the SIGSEGV handler would stay queued and not disrupt the
generation of the corefile.  At worst, corefile generation would take
above "shutdown_time" seconds (60), and OpenSIPS would SIGKILL its
entire process group and definitively terminate.

Now, OpenSIPS behavior has changed to:
    * attendant broadcasts a graceful IPC shutdown job
    * attendant gives _everyone_ a grand total of 5 seconds to finish
    * if not ready -> SIGKILL the entire instance

This behavior is too restrictive for the average spinning disk
performance.  Coupled with the fact that SHM pools of 2+ GB are common
nowadays AND the fact that, occasionally, corefiles come in multiples, 5
seconds is not a lot of time.

This patch fixes the behavior to:
    * attendant broadcasts a graceful IPC shutdown job
    * attendant gives non-crashed processes 5 seconds to obey
    * if not ready -> SIGKILL any processes which are still running
    * perform cleanup(60 sec) (leave the core-dumping processes alone)
    * exit or abort cleanup (leave the core-dumping processes alone)

Thanks to Bogdan for significantly directing the design of this patch