systemd: Properly shut down OpenSIPS 2.4+ instances
OpenSIPS 2.4 is meant to be shut down gracefully, in order to minimize
any form of data corruption caused by partial processing of SIP
messages.  Running "killall opensips" has a good chance of deadlocking
OpenSIPS for the next 45 seconds, especially if it's doing a lot of
logging.  This is often the case with the current systemd logic.

This patch updates the opensips systemd service files so they only
deliver SIGTERM to the attendant process.