dialog: Properly store all downstream forked legs
Commit a03ed4e1a36 over-simplified the downstream leg identification,
assuming that there is a 1:1 correspondence between outgoing branches
and unique replies to them.  This is not the case: as a proxy, we may
receive multiple responses (read: "To" tags) for each individual branch
we fork out.  This misconception would also cause SHM to leak.

We fix the above limitation by cloning a new dialog leg each time we
detect downstream forking of our outgoing branches.

Thanks to Bogdan for pointing out this problem