Added new function move_branch_to_ruri().
This move a branch (give by index) from the dset into the RURI & co.

This is part of the larger work to add support for dynamic following of the 3xx replies during a parallel forking attempt. Instead of waiting for the whole transaction to end (all branches), if a branch gets a 3xx reply, the redirect contact will be followed on spot by dynamically creating new branch (even if the transaction is not yet completed).

This is work sponsored by the Harris Corporation (https://www.harris.com)