Allow t_inject_branches("msg") to be used from onreply route.
In the context of the SIP reply, t_inject_branches() may be used to push as dynamic branches the destinations collected (via the dset) from the reply (like when the reply is a 3xx).

This is part of the larger work to add support for dynamic following of the 3xx replies during a parallel forking attempt. Instead of waiting for the whole transaction to end (all branches), if a branch gets a 3xx reply, the redirect contact will be followed on spot by dynamically creating new branch (even if the transaction is not yet completed).

This is work sponsored by the Harris Corporation (https://www.harris.com)