Fixed pass-through multi-part body with a single part
Previously the code was trying to convert the multi-part with a single part to a uni-part body, but this was incomplet as the Content-type was not updated and the per-part headers were lost.
So, a better approach is to keep it as it is (as packing).

(cherry picked from commit 953490775b82f2244eb3b0bcba6232d3d2af4484)