dialog: fix unset_dlg_profile() linkers deletion
This commit fixes the following bugs introduced in commit 00cddf7:
1. when it was called, it was removing all the linkers of a dialog
2. it was not removing the actual linker it was called for, causing
constant increase in size for that specific profile (and leak)
3. when temporary copying the profiles list, the values were not copied,
thus they might have dissapeared when the actual cleanup was done.

Many thanks to Jonathan Hulme for reporting the issue and providing
testing framework for detecting the bug and testing the fix!