async: fix fd leaking and invalid memory overwrite
when calling `reactor_del_reader(*fd, ...)`, the `*fd` will become -1.
The reason is because `fd` points to a `io_wait_h->fd_no`, that will get
changed to `-1` at the end of the function. Hence the actual `*fd` value
will be lost, and `close(*fd)` is a no-op, resulting in original fd
leaking. Moreover, when assigning `*fd` a new value, we will be
overwriting the `io_wait_h->fd_no` in the reactor, resulting in
undefined behavior.
This commit stores the actual value of `*fd` in a local variable and
uses it throught the function, making sure the `io_wait_h->fd_no` is not
overwritten and the `close()` function is called on the proper fd.

Many thangs to Eric Tamme (lirakis) on IRC for reporting this.