acc: fix leak when registering missed 2nd time
When calling do_accounting() twice, and the second time was containing
the "missed" flag, the actual acc context was leaking due to an extra
ref. This commit unrefs the context at the end of transaction.

Many thanks to Arto Kuiri for reporting this and for providing valuable
debugging information!

Close #1489