Fixed missing etag condition in DB query.
After refactoring the code to enable clustering in 2.4, some of the DB queries for updating / deleting the presentity records were missing the etag condition. As a result you may have experienced records leaking in DB or strange errors about unique key violation.
Many thanks for helping with the troubleshooting to Dan Pascu @ AG Projects and to Ognjen Šešlija.

Closes #1422
Closes #1378