usrloc: use contact id in update replicated msg
This commit also sends the contact_id in a replicated update packet.
This is useful when the instance that receives the replicated update
does not have the actual contact (it previously expired), and needs to
treat the update as an insert - before this patch insert would have
generated a new contact_id, possibly overlapping with another one
generated by the primary server.