Added clustering support (work in progress)
Each registation may have a sharing tag mapped. The actual registration will be performed only by the node (in the cluster) having that sharing tag as active.
The module will listen for all the changes of the sharing tag status, to see when the local node becomes active for one of its own tags -> if so, it must fire out all the registrations marked with that tag.
Still to be done: the unregister all and register when becoming active.