Same cleanup and re-organizing
The dialog ID is composed of call-id and an extra number -> prepare the groud for per-branch dialog-info sessions.

Register into TM module so we can have access to the incoming replies also (so we can publish the correct callee status, if multiple dialog-info sessions)

Pass the flags as parameters of the callback, instead of dlg_vals - much simpler.

Removed useless memset (are done by parse_to() function)

Work in progress for adding serial & parallel forking support in pua_dialoginfo