Fixed multiple terminates on the same process.
Mark the process to avoid counting it for other terminations. A process may continue to run after receiving the terminate RPC (due draining procedure), so it will continue to exist as running in the process table.
Use the OSS_PROC_TO_TERMINATE flag to avoid this (note that the flag is set and tested only by the attendent process).