DEV-1: Digest the pp tokens within lexer, not parser!
Using the lexer to consume the preprocessor tokens makes life so much
easier, as otherwise the complexity of the scripting language grammer
would skyrocket.  Example problem which is avoided by this:

if
<pp_token may go here>
(
<pp_token may go here>
condition
<pp_token may go here>
)
<pp_token may go here>
{
<pp_token may go here>
$var(x) = 5;
<pp_token may go here>
}

Although the coding style is horrible, this is still valid script
syntax, but, all of a sudden, the grammar must be extended in order to
support random tokens on any possible line of this "if" statement.
Nope, not going there.  Just use the lexer to eat the tokens and be done
with the problem!