Added "xlog_level" global parameter
This acts as log_level, but for the logs generated via xlog. This gives you separately control the verbosity level for the xlog() calls.
Added possibility to globally change the xlog_level during runtime (to be hooked later into an MI function) or to change the value per process (to be hooked later to the $xlog_level variable)