Fix t_reply_with_body() when called after t_newtran()
Force an update of the cloned request (from transaction) with info from the real request before internally running t_reply_with_body(). As t_reply_with_body() works with the cloned request, we need to be sure it is up-2-date; and t_newtran() does a partial cloning of the requests - so, if the transaction already exists better update its cloned request to be sure it has all data.
Credits for detailed reporting to @rrb3942
Closes #1632