replace opensipsctl snippets with opensips-cli
This is just a rough update of all examples that contained opensipsctl,
but should cover most of the cases. However, if there are any broken
examples, they should be reported properly.
Note that since opensipsctl was not yet removed, modules are still
deploying opensipsctl-related files. These will be sorted out when
opensipsctl and osipsconsole will be completely dropped