mid_registrar: Fix poorly handled lifetime reductions
If the downstream registrar reduces the lifetime of the contacts we
advertise to it, we (the mid-registrar) should properly adhere to the
change, and not utterly ignore it.

This bug was only present in modes 0 and 1.

Thanks to Jeff Pyle for the report!

(cherry picked from commit e5fd28efa238b5d25b5ebbce81d0a6421b04f1b8)