siprec: fix documentation to align with b6f2f84baa
In commit b6f2f84baa58e15baee43879308458bf38165b64 the signature of the
`siprec_start_recording` function was changed - instead of specifying
the rtpproxy set, now the function receives the rtpproxy node used for
that call. This commit fixes the documentation for that.

Thanks go to solarmon(GitHub) for reporting this in ticket #1664