cgrates: handle inter-mixing of requests and events
In case when OpenSIPS sends a request, but before receiving the reply on
the same connection, it receives an event, we need to make sure the
reply is not marked as completed.
Before this fix, when the actual reply was coming back, we no longer had
a reply handler, therefore OpenSIPS was crashing.

Reported by Vasilios Tzanoudakis (vtzan on GitHub) in ticket #1713

(cherry picked from commit c90d251cef30fb6a991b859652a3ba7bcb32c37c)