Refactored presence/rls configuration for specifying the Contact URI
Removed the server_address modparam and replaced it with contact_user.
The server_address was used to specify the URI for the Contact header
and it was optional for presence but mandatory for rls.

The idea of manually locking the Contact URI to a fixed value that is
specified in the configuration is non-functional in any setup that
involves the proxy listening on more than 1 IP/port/transport combo.

In the new model, the only configurable part of the Contact URI is the
username part, that can be specified through the contact_user modparam,
while the rest of the elements of the URI are automatically determined
from the interface where the request was recevied or sent from.