dialog: run loaded callback for dialogs loaded from db
Fixes a bug introduced in commit 8c8f27f that was relying on the fact
that dialogs were loaded from db *before* registering the loaded
callback, due to the dependencies mechanism. Thus the loaded callback
was ran on the spot, when the it was registered.

However, this is not entirely correct, since there are modules (such as
acc module) that do not have dialog dependencies, but still register
LOADED dialogs - in these cases, if dialog was loaded (in script) after
the module itself, the callback was never run. This behavior was
amplified by commit 3ddae43, which was basically never loading the
dialogs before registering the callback, thus the loaded callbacks were
never ran.

The purpose of this commit is to fix this issue, reported in #1754

Thanks go to Liviu Chircu for reporting the issue and providing the fix

(cherry picked from commit 6638b06f3ffa1fbc8d19f0532879e2c6c77a76d0)