tm: check branches that should be cancelled under lock
This fix avoids race condition between a CANCEL message and replies that
are coming in the same time.

Thanks go to Tommy Brecher for reprting this and offering extensive logs
to debug the issue in ticket #1759.
Credits for the fix are shared with Bogdan Iancu for the brainstoring.

(cherry picked from commit 6e39385233543ee9583469c55d3d3d1738917fb7)