dialog replication: Prevent crashes due to differing packet versions
Commits 58dc435cb563 and 852e629e4700 changed the format of the dialog
binary data packets.  This would cause an immediate crash during a
rolling upgrade, since upon upgrading and restarting the backup node, it
would sync from or receive packets from a primary node running the older
version, with the previous data format.

This patch makes it so dialog packets which do not meet the expected
version are simply discarded, rather than being left to cause a crash.