CSV parser: Relax the RFC 4180 implementation
Do not implement RFC 4180 down to the letter (e.g. double-quotes are
ONLY allowed within a field if they are escaped using another
double-quote).  Rather, only follow this rule if a field _begins_ with a
quote.  This helps maintain backwards-compatibility with the previous
OpenSIPS CSV parser, and is also desirable in many situations (e.g.
parsing a WWW-Authenticate header).

Fixes #1794