matching_mode honored in cachedbs
Function cdb_add_ct_update ignores matching modes set on opensips.cfg, this result in an unexpected behaviour if the user sets matching_mode=0 since multiple records with same contact (but different callid) are stored in the cache database.
This fix checks, before getting the base64 hash, if matching mode is contact only, if yes, the callid string is not included in the resulted base64 hash.
If matching mode is CONTACT_CALLID, this function behaves like before, and the hash is calcuated using contact:callid.