dialog: fix double lock get on replication update
When a replicated packet was received due to a timeout update, the
dialog was refferenced with the lock already acquired, resulting in a
deadlock. This fix refferences the dialog in an "unsafe" way, without
trying to acquire the lock, since it is already acquired.

(cherry picked from commit 6b0da180ab8ee33ac0eb784127e7738b90b4517e)