path: Do not omit ";received=" when merging Path headers
If "enable_double_path" is on and the inbound and outbound sockets
match, the second Path header would not be added anymore (good!),
thus skipping the ";received=" parameter completely (bad!).

The above behavior is actually a regression, introduced in 71c7d75f66c

Many thanks to Pete Kelly and Bogdan Iancu for spotting the problem!