registrar: Fix a ";received=" Path parsing bug with save("v")
Commit 71c7d75 introduced a parsing bug causing the ";received=" info to
not get stored into usrloc, as well as the NAT flag.

If we're saving these contacts with save("v") instead of save("p1v"), we
cannot route to them anymore (the latter version fully stores the Path,
so the front-end LB is still able to route back to the devices).

Credits to John Burke for the accurate report

Fixes #1809