tm: dup in pkg the faked msg headers
before this commit, the faked reqests were having the headers fields
stored in shm_mem and shared with the tm request uas. Thus, during
concurrent accesses to one of the headers' `parsed` field, the first
process would have parsed the field in pkg memory, and the second
process would have found that header already parsed. However, it
couldn't read the other's process memory, resulting in an invalid
access.
Moving the headers in a private area for the faked request solves this
issue, as each process will have its own `parsed` field in pkg memory.

Credits go to Rizwan Syed and Jonathan Hulme from Connex CS for
reporting and supporting the testing and fixing of this issue.

This issue has also been reported in ticket #1808