acc: always try to load dialog module
In case the `cdr` flag is used as variable, we can not figure out during
startup parsing whether the dialog module should be loaded or not.
Therefore we always try to load the dialog module, and in case we can't
and the `cdr` flag is used, we trigger a warning.