[acc] added loose dependency to dialog module
We need to be sure that dialog module is init before acc (if dialog is loaded), to prevent any crashes when acc registers the DLGCB_LOADED callback in dialog (the registraion of the callback will try to run the callback on any pre-existing dialogs, so the hash table for the dialogs must be created/init).
It is a loose dependency as it does not require the dialog module to be loaded. But if loaded, it must be init before ACC module.

Fixes #1833

(cherry picked from commit 3c18dbfc198d7bece144e708762ea35299bdc8b5)
(cherry picked from commit 7ade34c578e9219e66472e213f8ae70ec915dfbb)