Loop and wait for multiple terminted processes.
If you have a bunch of exec() calls (translated into a bunch of additional processes created), upon their termiantion OpenSIPS will get a cascade of SIGCHLD signals - as the kernel does not guarantee the delivery of the signals (like mergining similar multiple signal triggers), we will not have a 1-to-1 between terminated processes and received signals.
So, let's do our best and upon a single SIGCHLD we will try to "catch" as many terminated processes as possible/available.

Closes #1773