event_virtual: improve the behaviour of the FAILOVER policy
A previously failed subscriber is now skipped for further notifications
for a configurable interval (via the new "failover_timeout" modparam).
This is useful when the actual event backend blocks for a considerable
amount of time when rasing the event. This commit greatly decreases the
overall amount of time OpeSIPS processes remain stuck when a subscriber
is unresponsive.

Related to #1836

(cherry picked from commit 4dde71c13aaf5860472273c81bed7fec44e48723)