cdb_add_ct_update uses the full contact line to calculate the hash in the mongodb structure, for example:
"_id" : "#mysubscribernumberhere#",
	...
	"contacts" : {
		"c2lwOmd3Kzg0X0dXX1hYWFhYWFhAMTcyLjI4LjcxLjM6NTA4MDt0cmFuc3BvcnQ9dWRwO2d3PTg0X0dXX1hYWFhYWFg=" : {
			"attr" : null,
			"sip_instance" : null,
			...
			"contact" : "sip:gw+84_GW_XXXXXXX@172.28.71.3:5080;transport=udp;gw=84_GW_XXXXXXX"
		}
	}
In the above example the hash is a base64 representation of the value of the contact key.
This is not what we want because if inside the contact there is something that always changes like for example ";rinstance=#somedata#", we always get different hashes for the same registering endpoint leading to duplicate mongodb entries.
What we really need is to get a SIP URI based hash using and algorithm similar to get_ucontact().
Using this approach we end up having a single mongodb entry for each contact based only on the SIP URI taken from the contact SIP HEADER.