Access the route's array after the IPC dispatch.
The array with the script routes is not available in all procs, so, if the event is triggered from a non-scripting proc (like MI), the array will be empty.
As a fix, use the array (get the action list based on route ID) only via IPC dispatching, when we know for sure that we are in a process with the valid array of routes.
Fixes #1877

(cherry picked from commit 436dd69da5fcdb6fbb42a57e758c9696b662d0ef)