presence/clustering: Keep presence state on all nodes without shared DB
Add the option to set cluster_federation_mode = 2, which does the
following:

- All PUBLISH/presence state is recorded on the presence node, even when
  no one is subscribed. PUBLISH is now accepted on any node for updated
  state.

This means that you can round-robin your PUBLISH instead of having to
choose a single presence endpoint. And it also means you can switch to
other nodes and get up-to-date state from there, when a node dies.

This is achieved without using a shared DB (which has its own
limitations obviously).