dialog: build the ACK with full route set for re-INVITE
When the dialog module was sending re-INVITEs, the ACK to those
re-INVITEs was generated by the TM, thus was lacking a valid Route set.
This commit prevents the TM from generating the ACK, having the dialog
module sending it, with the learned Route set.

Thanks go to 46Labs for reporting and testing this!