event_flatstore: improve handling of multiple subscriptions for the same file
Use the same internal shm struct corresponding to a file for each evi
subscription for that file. This fixes an issue with the 'evi_flat_rotate' MI
command not properly "rotating" multiple subscriptions to the new file. Besides,
a file is now opened only once regardless of how many subscribtions there are
for the same file.
Also, do not reuse an old 'evi_reply_sock' structure for a new subscription for
the same file (if the file corresponding to an old subscription was not yet
closed by all processes). This led to incorrect expire times for new
subscriptions.

Fixes #1933