b2b_entities: don't try to insert in db nonprintable param
The b2b_entities module expects that the parameter provided during
entity registrations to be a string. It stores this string in the
database. However, there are cases where modules (such as siprec) force
their parameter to be a pointer to a structure, therefore the string
becomes non-printable. This breaks certain entities insertion in the
database.
This commit prevents the module from inserting a key that's not
printable. In the case of siprec this is not a problem, since the module
does not rely on the value that's stored in the parameter for restart -
it rebuilds the value anyway.

This is a temorary fix - ideally, the parameter should be able to be
serialized before written in the database; although this is not possible
now, it will be possible starting with 3.1. Hence this temporary fix.